Fix crash on level complete
On level complete the spawners were destroyed to clear the enemies from
the play area, however the Shooter script on the defenders still tried
to access the Attacker objects which caused the crash.  To fix this, all
defender objects are destroyed beforehand too which has the effect of
completely clearing the play space, and preventing the bug.